+++ daisy_ui_components_site/storybook/components/drawer.story.exs
@@ -13,6 +13,7 @@ defmodule Storybook.Components.Drawer do
           class: "z-10",
           selector_id: "my-drawer"
         },
+        # TODO: add the menu component when ready
         slots: [
           """
           <:drawer_content>
@@ -48,6 +49,31 @@ defmodule Storybook.Components.Drawer do
           """,
           """
           <:drawer_side class="h-full absolute">
+            <ul class="menu bg-base-200 text-base-content min-h-full w-80 p-4">
+              <!-- Sidebar content here -->
+              <li><a>Sidebar Item 1</a></li>
+              <li><a>Sidebar Item 2</a></li>
+            </ul>
+          </:drawer_side>
+          """
+        ]
+      },
+      %Variation{
+        id: :drawer_on_the_right_side,
+        attributes: %{
+          class: "z-10",
+          end: true,
+          selector_id: "my-drawer-4"
+        },
+        slots: [
+          """
+          <:drawer_content>
+            <!-- Page content here -->
+            <label for="my-drawer-4" class="drawer-button btn btn-primary">Open drawer</label>
+          </:drawer_content>
+          """,
+          """
+          <:drawer_side>
             <ul class="menu bg-base-200 text-base-content min-h-full w-80 p-4">
               <!-- Sidebar content here -->
               <li><a>Sidebar Item 1</a></li>