+++ lib/daisy_ui_components/join.ex
@@ -21,7 +21,7 @@ defmodule DaisyUIComponents.Join do
       </.join>
   """
   attr :class, :string, default: nil
-  attr :direction, :string, values: directions()
+  attr :direction, :string, values: ["horizontal", "vertical"]
   attr :rest, :global
   slot :inner_block, required: true