+++ CHANGELOG.md
@@ -5,7 +5,11 @@
 ### Added
 
 - Drawer Storybook
+- Navbar Storybook
 
 ### Changed
 
-- Drawer component with better naming convention on slots
+- Drawer with better naming convention on slots
+- Dropdown `align` property
+- Navbar slots for start, center and end
+- Documentation improvements